# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/brand/workspaces/{orgId}/map/violations": {
      "get": {
        "operationId": "BrandController_listViolations",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unacknowledgedOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/map/violations.csv": {
      "get": {
        "operationId": "BrandController_violationsCsv",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/authorized-retailers": {
      "get": {
        "operationId": "BrandController_listAuthorized",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      },
      "put": {
        "operationId": "BrandController_setAuthorized",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAuthorizedRetailersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/unauthorized-sellers": {
      "get": {
        "operationId": "BrandController_listUnauthorized",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/distribution/summary": {
      "get": {
        "operationId": "BrandController_distributionSummary",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/distribution": {
      "get": {
        "operationId": "BrandController_listDistribution",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "retailerSlug",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "sellerProductId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "belowMapOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/distribution/retailers": {
      "get": {
        "operationId": "BrandController_listRetailers",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/distribution/lapsed": {
      "get": {
        "operationId": "BrandController_listLapsed",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listings we have stopped seeing — reported as \"not seen since\", not \"delisted\".",
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/watches": {
      "get": {
        "operationId": "BrandController_listWatches",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      },
      "put": {
        "operationId": "BrandController_upsertWatch",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertBrandWatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/watches/{watchId}": {
      "delete": {
        "operationId": "BrandController_deleteWatch",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "watchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/discoveries": {
      "get": {
        "operationId": "BrandController_listDiscoveries",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/reports": {
      "get": {
        "operationId": "BrandController_listReports",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      },
      "post": {
        "operationId": "BrandController_generateReport",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Generate on demand. Idempotent for a given period.",
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/reports/{reportId}/csv": {
      "get": {
        "operationId": "BrandController_reportCsv",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/reports/{reportId}/pdf": {
      "get": {
        "operationId": "BrandController_reportPdf",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
    "/api/v1/brand/workspaces/{orgId}/reports/{reportId}": {
      "get": {
        "operationId": "BrandController_getReport",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brand"
        ]
      }
    },
...
  },
  "info": {
    "title": "PriceLens API",
    "description": "Cross-platform price comparison API: one canonical product per item, with every store listing matched to it and its price history.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SetAuthorizedRetailersDto": {
        "type": "object",
        "properties": {
          "platformIds": {
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "platformIds"
        ]
      },
      "UpsertBrandWatchDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "categoryId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "isOwnBrand": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "brand"
        ]
      },
      "GenerateReportDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ]
          }
        }
      },
...
    }
  }
}